track the byte length of heap allocations

+++ checker/src/k_limits.rs
@@ -10,7 +10,7 @@
 pub const MAX_ANALYSIS_TIME_FOR_BODY: u64 = 3;
 
 /// The maximum number of elements in a byte array that will be individually tracked.
-pub const MAX_BYTE_ARRAY_LENGTH: u128 = 10;
+pub const MAX_BYTE_ARRAY_LENGTH: usize = 10;
 
 /// Helps to limit the size of summaries.
 pub const MAX_INFERRED_PRECONDITIONS: usize = 50;
